move components into their own folders
