+++ itinerary_planner/src/main.jsx
@@ -1,9 +1,9 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
-import './index.css'
-import Home from './Home.jsx'
-import Login from './Login.jsx'
+import './styles/index.css'
+import Home from './Components/Home.jsx'
+import Login from './Components/Login.jsx'
 
 const router = createBrowserRouter([
   {
